nekodigi/streamlit3/requirements.txt
streamlit>=1.43
anthropic>=0.45
pillow>=11.0
